requirements.txt
langchain
langchain-core
langchain-openai
langchain-google-genai
langchain-community
python-dotenv
chromadb
tiktoken

